added register and login pages
